+++ openstack/cbr/v3/tasks/list.go
@@ -1,0 +1,37 @@
+package tasks
+
+import (
+	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
+)
+
+type ListOpts struct {
+	EndTime             string `q:"end_time"`
+	EnterpriseProjectId string `q:"enterprise_project_id"`
+	Limit               int    `q:"limit"`
+	Offset              int    `q:"offset"`
+	OperationType       string `q:"operation_type"`
+	ProviderId          string `q:"provider_id"`
+	ResourceId          string `q:"resource_id"`
+	ResourceName        string `q:"resource_name"`
+	StartTime           string `q:"start_time"`
+	Status              string `q:"status"`
+	VaultId             string `q:"vault_id"`
+	VaultName           string `q:"vault_name"`
+}
+
+func List(client *golangsdk.ServiceClient, opts ListOpts) ([]OperationLog, error) {
+	query, err := golangsdk.BuildQueryString(opts)
+	if err != nil {
+		return nil, err
+	}
+
+	raw, err := client.Get(client.ServiceURL("operation-logs")+query.String(), nil, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	var res []OperationLog
+	err = extract.IntoSlicePtr(raw.Body, &res, "operation_log")
+	return res, nil
+}
